+++ mutant-cli/src/callbacks/mod.rs
@@ -4,3 +4,4 @@ mod progress;
 pub mod put;
 
 pub use init::create_init_callback;
+pub use progress::StyledProgressBar;